`clippy_dev`: Separate the IR definitions into a separate module.
