+++ clippy_dev/src/generate.rs
@@ -1,5 +1,5 @@
+use crate::ir::{ActiveLint, ConfDef, LintData, LintPass, ParsedLints};
 use crate::parse::cursor::Cursor;
-use crate::parse::{ActiveLint, ConfDef, LintData, LintPass, ParsedLints};
 use crate::utils::{FileUpdater, UpdateMode, UpdateStatus, VecBuf, slice_groups, update_text_region_fn};
 use core::range::Range;
 use itertools::Itertools as _;
