+++ clippy_dev/src/lib.rs
@@ -26,6 +26,7 @@ extern crate termize;
 pub mod dogfood;
 pub mod edit_lints;
 pub mod fmt;
+pub mod ir;
 pub mod lint;
 pub mod new_lint;
 pub mod release;
